TIKA-3695 -- implement filterExisting in StandardWriteFilter.  Create standalone unit test.
