+++ tika-core/src/test/java/org/apache/tika/config/TikaConfigTest.java
@@ -23,11 +23,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
@@ -40,9 +38,6 @@ import org.apache.tika.ResourceLoggingClassLoader;
 import org.apache.tika.exception.TikaConfigException;
 import org.apache.tika.exception.TikaException;
 import org.apache.tika.metadata.Metadata;
-import org.apache.tika.metadata.MetadataWriteFilterFactory;
-import org.apache.tika.metadata.StandardWriteFilterFactory;
-import org.apache.tika.metadata.TikaCoreProperties;
 import org.apache.tika.mime.MediaType;
 import org.apache.tika.mime.MimeDetectionTest;
 import org.apache.tika.parser.AutoDetectParser;
@@ -51,7 +46,6 @@ import org.apache.tika.parser.CompositeParser;
 import org.apache.tika.parser.DefaultParser;
 import org.apache.tika.parser.EmptyParser;
 import org.apache.tika.parser.ErrorParser;
-import org.apache.tika.parser.ParseContext;
 import org.apache.tika.parser.Parser;
 import org.apache.tika.parser.ParserDecorator;
 import org.apache.tika.parser.mock.MockParser;
@@ -380,7 +374,6 @@ public class TikaConfigTest extends AbstractTikaConfigTest {
         });
     }
 
-
     @Test
     public void testTimesInitiated() throws Exception {
         //this prevents multi-threading tests, but we aren't doing that now...
@@ -400,67 +393,5 @@ public class TikaConfigTest extends AbstractTikaConfigTest {
         assertNull(config.getMaximumCompressionRatio());
         assertNull(config.getMaximumDepth());
         assertNull(config.getMaximumPackageEntryDepth());
-    }
-
-    @Test
-    public void testMetadataFactoryConfig() throws Exception {
-        TikaConfig tikaConfig =
-                new TikaConfig(TikaConfigTest.class.getResourceAsStream("TIKA-3695.xml"));
-        AutoDetectParserConfig config = tikaConfig.getAutoDetectParserConfig();
-        MetadataWriteFilterFactory factory = config.getMetadataWriteFilterFactory();
-        assertEquals(241, ((StandardWriteFilterFactory) factory).getMaxEstimatedBytes());
-        AutoDetectParser parser = new AutoDetectParser(tikaConfig);
-        String mock = "<?xml version=\"1.0\" encoding=\"UTF-8\" ?>" +
-                "<mock>";
-        for (int i = 0; i < 20; i++) {
-            mock += "<metadata action=\"add\" name=\"dc:creator\">01234567890123456789</metadata>";
-        }
-        mock += "<write element=\"p\" times=\"30\"> hello </write>\n";
-        mock += "</mock>";
-        Metadata metadata = new Metadata();
-        List<Metadata> metadataList =
-                getRecursiveMetadata(new ByteArrayInputStream(mock.getBytes(StandardCharsets.UTF_8)),
-                parser, metadata, new ParseContext(), true);
-        assertEquals(1, metadataList.size());
-        metadata = metadataList.get(0);
-
-        String[] creators = metadata.getValues("dc:creator");
-        assertEquals(9, creators.length);
-        assertEquals("0123456", creators[8]);
-        assertContainsCount(" hello ", metadata.get(TikaCoreProperties.TIKA_CONTENT), 30);
-        assertEquals("true", metadata.get(TikaCoreProperties.METADATA_LIMIT_REACHED));
-    }
-
-    @Test
-    public void testMetadataFactoryFieldsConfig() throws Exception {
-        TikaConfig tikaConfig =
-                new TikaConfig(TikaConfigTest.class.getResourceAsStream("TIKA-3695-fields.xml"));
-        AutoDetectParserConfig config = tikaConfig.getAutoDetectParserConfig();
-        MetadataWriteFilterFactory factory = config.getMetadataWriteFilterFactory();
-        assertEquals(241, ((StandardWriteFilterFactory) factory).getMaxEstimatedBytes());
-        AutoDetectParser parser = new AutoDetectParser(tikaConfig);
-        String mock = "<?xml version=\"1.0\" encoding=\"UTF-8\" ?>" +
-                "<mock>";
-        mock += "<metadata action=\"add\" name=\"dc:subject\">this is not a title</metadata>";
-        mock += "<metadata action=\"add\" name=\"dc:title\">this is a title</metadata>";
-        for (int i = 0; i < 20; i++) {
-            mock += "<metadata action=\"add\" name=\"dc:creator\">01234567890123456789</metadata>";
-        }
-        mock += "<write element=\"p\" times=\"30\"> hello </write>\n";
-        mock += "</mock>";
-        Metadata metadata = new Metadata();
-        List<Metadata> metadataList =
-                getRecursiveMetadata(new ByteArrayInputStream(mock.getBytes(StandardCharsets.UTF_8)),
-                        parser, metadata, new ParseContext(), true);
-        assertEquals(1, metadataList.size());
-        metadata = metadataList.get(0);
-
-        String[] creators = metadata.getValues("dc:creator");
-        assertNull(metadata.get("dc:subject"));
-        //this gets more than the other test because this is filtering out X-TIKA:Parsed-By"
-        assertEquals(12, creators.length);
-        assertEquals("012345", creators[11]);
-        assertContainsCount(" hello ", metadata.get(TikaCoreProperties.TIKA_CONTENT), 30);
-        assertEquals("true", metadata.get(TikaCoreProperties.METADATA_LIMIT_REACHED));
     }
 }
